Support non-webpack environments
This fix introduces a number of changes that enable Firebolt to work in projects that do not use WebPack:

- renamed all Firebolt source files to `.mjs` since they are all modules
- removed webpack.config.js and all usage of webpack
- added appropriate entry and types metadata to package.json
- change main entry `index.mjs` to `firebolt.mjs`
- updated all of the tests to import the new main entry point
- fixed a race condition in `Lifecycle/defaults.mjs` that occurs due to how modules load different than .js
