+++ package.json
@@ -2,8 +2,8 @@
   "name": "@firebolt-js/sdk",
   "version": "0.6.1-beta.2",
   "description": "The Firebolt JS SDK",
-  "main": "./dist/firebolt.mjs",
-  "types": "./dist/firebolt.d.ts",
+  "main": "./dist/lib/firebolt.mjs",
+  "types": "./dist/lib/firebolt.d.ts",
   "scripts": {
     "prepare": "npm run prepare:setup && npm run dist:docs && npm run dist:sdk && npm run dist:d",
     "prepare:setup": "mkdir -p ./dist/docs ./build/docs/markdown ./build/sdk/javascript/src",
@@ -12,14 +12,13 @@
     "build:sdk": "firebolt-openrpc --task sdk --source ./src/ --template ./src/template/js --output ./build/sdk/javascript/src --static-modules Platform --shared-schemas node_modules/@firebolt-js/schemas/src/schemas/",
     "build:docs": "firebolt-openrpc --task docs --source ./src/ --template ./src/template/markdown --output ./build/docs/markdown --shared-schemas node_modules/@firebolt-js/schemas/src/schemas/ --as-path",
     "build:wiki": "firebolt-openrpc --task docs --source ./src/ --template ./src/template/markdown --output ./build/docs/markdown --shared-schemas node_modules/@firebolt-js/schemas/src/schemas/",
-    "dist:sdk": "npm run validate && npm run build:sdk && npm run pack && npm run test",
-    "dist:d": "npm run validate && firebolt-openrpc --task declarations --source ./src/ --output ./dist/firebolt.d.ts --shared-schemas node_modules/@firebolt-js/schemas/src/schemas/",
+    "dist:sdk": "npm run validate && npm run build:sdk && cp -R ./build/sdk/javascript/src ./dist/lib && npm run test",
+    "dist:d": "npm run validate && firebolt-openrpc --task declarations --source ./src/ --output ./dist/lib/firebolt.d.ts --shared-schemas node_modules/@firebolt-js/schemas/src/schemas/",
     "dist:docs": "npm run validate && npm run build:docs asPath && rm -r ./dist/docs && cp -R ./build/docs/markdown ./dist/docs",
     "dist:wiki": "npm run validate && npm run build:wiki && rm -r ./dist/docs && cp -R ./build/docs/markdown ./dist/docs",
     "test": "jest --config=jest.config.js --detectOpenHandles",
-    "clean": "rm -rf build && mkdir -p build",
+    "clean": "rm -rf build && mkdir -p build && rm -rf dist && mkdir -p dist",
     "build": "npm run validate && npm run build:docs && npm run build:sdk",
-    "pack": "webpack --config ./webpack.config.js",
     "publish:docs": "npm run build:docs && sh ./push-docs.sh",
     "publish:word": "mkdir -p ./dist/word/schemas && cd ./dist/docs && find . -iname \"*.md\" -type f -exec sh -c 'pandoc \"${0}\" -o \"../word/${0%.md}.docx\"' {} \\;"
   },
@@ -36,12 +35,10 @@
     "@babel/core": "^7.14.6",
     "@babel/node": "^7.14.5",
     "@babel/preset-env": "^7.14.5",
-    "@firebolt-js/openrpc": "1.3.0-alpha.1",
+    "@firebolt-js/openrpc": "../firebolt-openrpc",
+    "@firebolt-js/schemas": "0.2.0-alpha.1",
     "ajv": "^6.12.6",
-    "jest": "^27.0.4",
-    "webpack": "^5.39.1",
-    "webpack-cli": "^4.2.0",
-    "@firebolt-js/schemas": "0.2.0-alpha.1"
+    "jest": "^27.0.4"
   },
   "keywords": [
     "firebolt",
