+++ src/template/js/sdk/Platform/defaults.mjs
@@ -16,46 +16,12 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import Transport from '../Transport'
-/* ${IMPORTS} */
+import accessibility from '../Accessibility/defaults.mjs'
+import device from '../Device/defaults.mjs'
+import localization from '../Localization/defaults.mjs'
 
-/* ${INITIALIZATION} */
-
-export const store = {
-  _current: 'initializing',
-  get current() {
-    return this._current
-  }
-}
-
-Events.listen('Lifecycle', (event, value) => {
-  store._current = event
-})
-
-/* ${METHODS} */
-
-function state() {
-  return store.current
-}
-
-function finished() {
-  if (store.current === 'unloading') {
-    return Transport.send('lifecycle', 'finished')
-  } else {
-    throw 'Cannot call finished() except when in the unloading transition'
-  }
-}
-
-// public API
 export default {
-
-  /* ${EVENTS} */
-
-  /* ${ENUMS} */
-
-  state,
-  finished,
-
-  /* ${METHOD_LIST} */
-  
+  localization: localization,
+  device: device,
+  accessibility: accessibility,
 }
