+++ test/suite/discovery.test.js
@@ -17,7 +17,7 @@
  */
 
 import Setup from '../Setup'
-import { Discovery } from '../../dist/firebolt.mjs'
+import { Discovery } from '../../dist/lib/firebolt.mjs'
 
 test('watched(contentId)', () => {
     return Discovery.watched('abc').then(success => {
