+++ test/suite/lifecycle.test.js
@@ -17,7 +17,7 @@
  */
 
 import { testHarness } from '../Setup'
-import { Lifecycle } from '../../dist/firebolt.mjs'
+import { Lifecycle } from '../../dist/lib/firebolt.mjs'
 
 let readyResolved = false
 let readyCalled = false
