+++ test/suite/metrics.test.js
@@ -17,7 +17,7 @@
  */
 
 import Setup from '../Setup'
-import { Metrics } from '../../dist/firebolt.mjs'
+import { Metrics } from '../../dist/lib/firebolt.mjs'
 
 test('Sign-in / Sign-out', () => {
     Metrics.startContent('abc')
